Update cinepak/cpkdemo to use Makefile helpers
Similar to the cinepak/sample update, use
jagdefs.mk and jagrules.mk to build the more
complex cinepak demo/example.  Again, remove
cpkdemo.lnk because it just added another place
where the object file list had to be duplicated.

This is untested for now, but it works well enough
to show the "Licensed to Atari" splash screen.