Update to working 3D library demo
The 3D demo from Belboz's SDK was claimed to never
have been working.  Replace it with a newer,
working copy.  The prior readme.txt dated the code
April 25, 1995.  This version is dated
September 14, 1995, and does in fact work.  It was
taken from the "RENDER" directory in the file
Development.zip available on this thread:

http://www.3do.cdinteractive.co.uk/viewtopic.php?f=35&t=3430

Then modified as follows:

-Everything to lowercase, capitalize Makefile

-chmod -x *

-Delete object files and binaries

-dos2unix on the remaining files

-Fixed lib/font.c to not truncate blitter flags

-Integrate the Makefile with the updated SDK build
 framework.

-Some manual merges with the old code:

 * Use the union of the models/textures available
   in both the old and new 3d demo.

 * Hand-edit the hand-rolled "cube" (really a two-
   sided rectangle) model from the old demo to get
   it to compile with the new code.

 * Preserve the hand-coded checkerboard textures
   in miscasm.s.