Use dosemu for most DOS-only tools on Linux
Add a dosemu_helper script that invokes a Jaguar
SDK DOS command in a dosemu session on Linux.  It
takes care of ensuring the filesystem is set up to
access the command and any files passed as
parameters on the command line.  Still to do:

aln and mac - Not usable directly on Linux anymore,
              but they need a little more work to
              ensure their include paths and
              whatnot are forwarded correctly.

wdb and rdbjag - These likely aren't useful inside
                 a dosemu session anyway, nor are
                 they likely to be much use to
                 most developers anyway these
                 days.