Add commented support for mac/aln in jagdefs.mk
It's possible to substitute the original
brainstorm mac/aln Jaguar assembler/linker for
rmac and rln.  This is useful for testing and
comparison when something fails to build with
rmac/rln.

Currently, this only works on systems that support
a.out binaries, but hopefully optional dosemu
wrapper support will be added soon :-)