skunk lib: skunkFILEREAD return value now fills d0
Prior to this change, only a word of d0 was filled
with the return value, and the upper two bytes may
have had stale data from the timeout counter in
them.  This was fine if the caller didn't check
for errors or only used the lower word, but broke
when using cmp.l on the return value.