Use Kees Cook's a.out loader to run a.out binaries
This little tool from the kernel-tools repository
is an ELF32 program that essentially loads a
statically linked a.out binary into memory and
executes it, avoiding the need for an in-kernel
a.out binary loader. This allows the Linux kernel
to proceed with removal of its x86 a.out loader
without breaking the use of these ancient Atari
binaries on Linux.