Use GPU/DSP register dump/set code from rdbjag
This register dump/set code, extracted by running
wdb and disassembling everything below 0x2000,
clobbers one less register than the prior version
by letting the debugger take care of stopping the
JRISC processor. This may not be completely
reliable when interrupt handlers are in use on the
processor in question, but we could probably force
those off while running this code if it becomes an
issue in practice.

This code is also significantly shorter than the
prior code.