Add GPU go-until-break command: gob
This command, ported from gpu.db, is similar to
the GDB command "advance:" It sets a breakpoint at
the specified location and then runs the GPU at
full speed until that breakpoint is hit, at which
the GPU is paused. The user can then examine the
GPU registers, continue stepping through the GPU
code, or set another breakpoint and run to it.