Modify dhrystone-gpu example to build in SDK
Good example of how easy it is to add GPU code
written in C to your SDK-based project.

Run "make" to build, and run it on a skunkboard
with "jcp -c dhry.cof". It doesn't display
anything, but if you're patient, it'll print the
benchmark results to the skunk console in about
10 seconds.