Update URL for config.sub/config.guess in maketools.sh
The old URLs have been replaced with newer ones,
and the redirect throughseems to be failing quite
often, and gitweb itself doesn't seem to work
reliably. This cgit one isn't terribly good
either, 2/3 of the IPs it resolves to are
currently timing out, but at least the the 3rd one
works, and presumably the others will be fixed or
removed from the pool eventually.